Delete the qt interface
Summary:
The qt wallet system is a significant amount of code to maintain.  In particular the build system for the qt binary is very sensitive.  As the goal of this project is to enable third party wallets, it would be best if people use some other wallet to maintain their funds, or a system which uses the RPC interfaces.

Test Plan:
Verify building still works

Tags: #purge